requirements.txt
APScheduler==3.7.0
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
cloudinary==1.36.0
dj-database-url==2.1.0
dj3-cloudinary-storage==0.0.6
Django==4.1
django-timezone-field==6.1.0
django-use-email-as-username==1.4.0
feature-engine==1.6.2
gunicorn==21.2.0
idna==3.6
joblib==1.3.2
numpy==1.26.2
packaging==23.2
pandas==2.1.4
patsy==0.5.5
psycopg2==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sqlparse==0.4.4
statsmodels==0.14.1
threadpoolctl==3.2.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==2.1
urllib3==2.1.0
whitenoise==6.6.0
djangorestframework==3.15.1
django-cors-header==4.3.1
# xgboost==2.0.3
